Polar

AirfoilPolars.Polar — Type

Polar

A struct to store airfoil polars.

Fields

Re::Float64: Reynolds number

alpha::Vector{Float64}: angle of attack [deg]

cl::Vector{Float64}: lift coefficient

cd::Vector{Float64}: drag coefficient

cm::Vector{Float64}: moment coefficient

M::Float64: Mach number

n_crit::Float64: critical amplification factor

xtrip::Tuple{Float64,Float64}: transition location

name_airfoil::String: name of the airfoil

AirfoilPolars.smooth — Function

smooth(
    p::Polar; 
    order_cl=3, order_cd=3, order_cm=3, 
    smoothing_cl=0.01, smoothing_cd=0.0001, smoothing_cm=0.005
)

Smooths a polar using splines.

AirfoilPolars.blend — Function

blend(p1::Polar, p2::Polar, weight2::Float64)

Blends two polars using a weight.

Arguments

p1::Polar: first polar

p2::Polar: second polar

weight2::Float64: weight for the second polar

AirfoilPolars.extrapolate — Function

extrapolate(
    p::Polar; 
    cd_max::Union{Nothing,Float64}=nothing, 
    AR::Union{Nothing,Float64}=nothing, 
    cd_min=0.001, 
    n_alpha::Int=15
)

Extrapolate a polar to +/- 180 degrees. This function is based on the Viterna method.

Arguments

p::Polar: polar to extrapolate

cd_max::Union{Nothing,Float64}: maximum drag coefficient

AR::Union{Nothing,Float64}: aspect ratio

cd_min::Float64: minimum drag coefficient

n_alpha::Int: number of points to use for extrapolation

Corrections

AirfoilPolars.correction3D — Function

correction3D(
    polar::Polar, 
    r::Float64, c::Float64, u_inf::Float64, Ω::Float64; 
    alpha_max_corr::Float64=30.0, alpha_linear_min::Float64=-5.0, alpha_linear_max::Float64=5.0
)

Applies 3-D corrections for rotating sections from the 2-D data using the Du and Selig model.

Parameters

polar::Polar: 2-D polar

r::Float64: local radius

c::Float64: local chord

u_inf::Float64: local airfoil freestream velocity ≈ √(Vwind^2 + (Ωr)^2)

Ω::Float64: rotation speed [rad/s]

alpha_max_corr::Float64: maximum angle of attack to apply full correction

alpha_linear_min::Float64: angle of attack where linear portion of lift curve slope begins

alpha_linear_max::Float64: angle of attack where linear portion of lift curve slope ends

Returns

polar::Polar: A new Polar object corrected for 3-D effects

References

Z. Du and M. Selig, ‘A 3-D stall-delay model for horizontal axis wind turbine performance prediction’, in 1998 ASME Wind Energy Symposium, Reno, NV, U.S.A.: American Institute of Aeronautics and Astronautics, Jan. 1998. doi: 10.2514/6.1998-21.

AirfoilPolars.correction_Mach — Function

correction_Mach(polar::Polar, Ma_new)

Corrects a polar for a new Mach number using Prandtl-Glauert compressibility correction:

$c_l = \frac{c_{l, M=0}}{\sqrt{1 - M^2}}$

Miscalleneous

AirfoilPolars.generate_name — Function

generate_name(p::Polar; fname_extra="")

Generates a name for a polar based on its properties.

Arguments

p::Polar: polar to generate name for

fname_extra::String: extra string to append to the name

AirfoilPolars.plot — Function

plot(polars::Vector{Polar}; fname=nothing, dpi=300, legend=true)

Plots a vector of polars.