Solve

AirfoilPolars.init — Function

init(af::Airfoil)

Initialize the airfoil for the Xfoil solver.

AirfoilPolars.solve — Function

solve(
    alpha::Vector, re::Number; 
    n_iter::Int = 100, n_crit::Number = 9.0, 
    xtrip::Tuple = (1.0, 1.0), 
    mach::Number = 0.0, 
    make_nonconverged_nan::Bool = false, 
    interpolate_nonconverged::Bool = false
)

Solve the airfoil at different angles of attack and Reynolds numbers using Xfoil.

Arguments

alpha::Vector: vector of angles of attack

re::Number: Reynolds number

n_iter::Int: number of iterations

n_crit::Number: critical value of the transition criterion

xtrip::Tuple: transition location

mach::Number: Mach number

make_nonconverged_nan::Bool: if true, the values of cl, cd, and cm for non-converged cases are set to NaN

interpolate_nonconverged::Bool: if true, the values of cl, cd, and cm for non-converged cases are interpolated